builder instead of create!
user class has now some strings as input

To supply firstName and lastName we have to create an interface inside our UserComponent interface and annotate it with @Component.Builder. Whenever we create a Builder interface like this we always have to create a method that returns the component type (UserComponent here).